reorganize code + make sure file are deleted
